pyproject.toml
[tool.poetry]
name = "customer-support-chat"
version = "0.1.0"
description = "Customer Support Chat Application"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.2.19"
python-dotenv = "^1.0.1"
ipython = "^8.27.0"
duckduckgo-search = "^6.2.11"
langchain-community = "^0.2.16"
langchain-google-community = "^1.0.8"
qdrant-client = "^1.11.1"
pandas = "^2.2.2"
aiohttp = "^3.10.5"
tqdm = "^4.66.5"
more-itertools = "^10.5.0"

# Optional: remove if no longer using OpenAI
# langchain-openai = "^0.1.23"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
